Simplify data-loading delegating neighbor calculations to backbone wrappers
