+++ franken/data/__init__.py
@@ -1,8 +1,8 @@
-from franken.data.base import BaseAtomsDataset, Configuration, Target
-
+from franken.data.base import Configuration, Target
+from franken.data.dataset import FrankenAtomsDataset
 
 __all__ = [
-    "BaseAtomsDataset",
     "Configuration",
     "Target",
+    "FrankenAtomsDataset",
 ]
